Now can handle Selective Dynamics Tags in POSCAR
A very basic edit to the code, which allows the new code to be used with Selective Dynamics tags in POSCAR. Given accurate POSCAR in the first image and final image. NEB with selective dynamics are very helpful in systems with a large number of atoms, and we don't want all of them to move.